requirements.txt
fastapi
uvicorn
python-dotenv
pytest
requests
properties